+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `source` field in `tooth.json`.
+
 ## [0.19.0] - 2024-01-20
 
 ### Added